only do this for BD

+++ db/data/20250204101724_cancel_bangladesh_stale_experiments.rb
@@ -2,8 +2,9 @@
 
 class CancelBangladeshStaleExperiments < ActiveRecord::Migration[6.1]
   def up
-    Experimentation::Experiment.upcoming.where(experiment_type: 'stale_patients').each do |experiment|
-      experiment.destroy  # ...because we need the associated notifications gone too
+    return unless CountryConfig.current_country?("Bangladesh") && SimpleServer.env.production?
+    Experimentation::Experiment.upcoming.where(experiment_type: "stale_patients").each do |experiment|
+      experiment.destroy # ...because we need the associated notifications gone too
     end
   end
